+++ src/domains/services/generator.service.ts
@@ -7,7 +7,6 @@ import { Specification } from '../models/SpecificationFile';
 import { BaseService } from './base.service';
 
 import AsyncAPIGenerator from '@asyncapi/generator';
-import AsyncAPINewGenerator from 'generator-v2';
 import { spinner } from '@clack/prompts';
 import path from 'path';
 import os from 'os';
@@ -76,6 +75,7 @@ export class GeneratorService extends BaseService {
     if (v3NotSupported) {
       return this.createErrorResult(v3NotSupported);
     }
+    const logs: string[] = [];
 
     const generator = new AsyncAPIGenerator(
       template,
@@ -84,7 +84,7 @@ export class GeneratorService extends BaseService {
     );
     const s = interactive
       ? spinner()
-      : { start: () => null, stop: (string: string) => console.log(string) };
+      : { start: () => null, stop: (string: string) => logs.push(string) };
     s.start('Generation in progress. Keep calm and wait a bit');
     try {
       await generator.generateFromString(asyncapi.text(), {
@@ -92,48 +92,10 @@ export class GeneratorService extends BaseService {
         path: asyncapi,
       });
     } catch (err: any) {
-      console.log(err);
       s.stop('Generation failed');
       return this.createErrorResult(err.message, err.diagnostics);
     }
     s.stop(
-      this.getGenerationSuccessMessage(output),
-    );
-
-    return this.createSuccessResult({
-      success: true,
-      outputPath: output,
-    } as GenerationResult);
-  }
-
-  async generateUsingNewGenerator(
-    asyncapi: Specification,
-    template: string,
-    output: string,
-    options: any,
-    genOption: any,
-  ): Promise<ServiceResult<GenerationResult>> {
-    const v3NotSupported = this.checkV3NotSupported(asyncapi, template);
-    if (v3NotSupported) {
-      return this.createErrorResult(v3NotSupported);
-    }
-    const logs = [];
-
-    const generator = new AsyncAPINewGenerator(
-      template,
-      output || path.resolve(os.tmpdir(), 'asyncapi-generator'),
-      options,
-    );
-    try {
-      await generator.generateFromString(asyncapi.text(), {
-        ...genOption,
-        path: asyncapi,
-      });
-    } catch (err: any) {
-      logs.push('Generation failed');
-      return this.createErrorResult(err.message, err.diagnostics);
-    }
-    logs.push(
       this.getGenerationSuccessMessage(output),
     );